Update to match Jorge's suggestions. Initial search outline
